# src/swagger/swagger.yaml
swagger: '2.0'
host: 'localhost:1234'
basePath: /
schemes:
  - http
info:
  title: Theater ticket API Documentation
  version: 0.0.1
  contact:
    name: Karun
    email: karunpmenon@gmail.com
paths:
  /deletewithFilter:
    delete:
      summary: delete docs based on customerName
      operationId: deletewithfilter
      description: delete docs based on filter
      parameters:
        - name: customerName
          in: body
          description: name of the customer to be deleted
          schema:
            type: string
          required: true
      responses:
        default:
          schema:
            type: string
          description: Successful
...